Cierra las cinco fases de la auditoria de backend (docs/BACKEND-AUDIT.md)
y los cuatro hallazgos de infra que quedaban de la de seguridad
(docs/SECURITY-AUDIT.md). Suite de 204 a 304 tests.

BACKEND — 5 fases:
  1. Indices, fetch joins y paginacion. Encontro un bug que nadie habia
     reportado: el panel del admin NO mostraba ningun ticket de
     herramienta (INNER JOIN implicito de la Criteria API).
  2. I/O de red fuera de las transacciones (OCR y timeouts de MinIO).
  3. Cobertura de las clases sin tests. Encontro un segundo bug: la
     validacion del alta masiva devolvia 500 en vez de 400.
  4. Eficiencia: proyecciones, indice cubridor, y filtrar en la base lo
     que se filtraba en memoria.
  5. litros de Double a BigDecimal (DB-04), con migracion V12.

SEGURIDAD — 11 de 12:
  SEC-07 y SEC-08 en nginx, SEC-12 con el registro de eventos.
  SEC-11 (mem_limit) queda ABIERTO por decision explicita: sin conocer la
  RAM del VPS, un limite mal puesto mata el contenedor. Riesgo residual
  documentado en el hallazgo.

DEPLOY: requiere correr las migraciones V10, V11 y V12, y sacar APK nuevo
por el cambio de /tickets/me a PagedModel (breaking, commit 08b2b6e).

+++ backend/src/main/java/com/retrorental/backend/dto/response/ProveedorConsumo.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 public record ProveedorConsumo(
     Integer proveedorId,
     String nombre,
-    double litros,
+    BigDecimal litros,
     BigDecimal gasto
 ) {
 }
